re-organized desktop to host and mobile to client in server code. Set stage to check for nonce in url param and query as options. Added parameter for handlePost to configure where in the body the object will be saved.
