+++ lib/server/clientController.js
@@ -1,7 +1,7 @@
 "use strict"; // eslint-disable-line
 /* eslint-disable no-param-reassign */
 const jwtController = require('./jwtController.js');
-const mobileController = {};
+const clientController = {};
 
 /**
  * If a token exists, extract the room id from the token and store it on cookie
@@ -10,7 +10,7 @@ const mobileController = {};
  * @param {Object} res - response object
  * @param {Object} connectRequests - a reference to the activeConnectRequests object
  */
-mobileController.handleGet = (req, res, connectRequests) => {
+clientController.handleGet = (req, res, connectRequests) => {
   if (req.useragent && req.useragent.isMobile) {
     // console.log(`request is from Mobile`);
     // check for token / room session
@@ -25,7 +25,7 @@ mobileController.handleGet = (req, res, connectRequests) => {
 };
 
 /**
- * Verifies nonce sent in the req body under key 'codeCheck'
+ * Verifies nonce sent in the req body a user configurable bodyTag
  * If nonce is valid, set corresponding room id to a cookie on response
  *   (this will initiate a socket connection when the browser gets the response)
  *   then sets req.imperio's connected property to true
@@ -33,23 +33,20 @@ mobileController.handleGet = (req, res, connectRequests) => {
  * @param {Object} res - response object
  * @param {Object} connectRequests - a reference to the activeConnectRequests object
  */
-mobileController.handlePost = (req, res, connectRequests) => {
-  // console.log('mobile handlePost called');
-  if (req.useragent && req.useragent.isMobile) {
-    // console.log('post is from Mobile: ', req.body);
-    // if correct then redirect to tap page
-    // TODO how do we allow the developer to link up
-    //   their input that sends the nonce with our module?
-    const nonce = req.body.codeCheck;
-    if (connectRequests.hasOwnProperty(nonce)) {
-      const roomId = connectRequests[nonce].roomId;
-      // console.log('the nonce exists and matches to roomId:', roomId);
-      jwtController.createTokenFrom(roomId, res);
-      res.cookie('roomId', roomId); // DEBUG this is for reference?
-      req.imperio.connected = true;
-    }
-    // if incorrect then redirect to rootmobile page with error message
+clientController.handlePost = (req, res, connectRequests, bodyTag) => {
+  // console.log('client handlePost called');
+  // if correct then redirect to tap page
+  // TODO how do we allow the developer to link up
+  //   their input that sends the nonce with our module?
+  const nonce = req.body[bodyTag];
+  if (connectRequests.hasOwnProperty(nonce)) {
+    const roomId = connectRequests[nonce].roomId;
+    // console.log('the nonce exists and matches to roomId:', roomId);
+    jwtController.createTokenFrom(roomId, res);
+    res.cookie('roomId', roomId); // DEBUG this is for reference?
+    req.imperio.connected = true;
   }
+    // if incorrect then redirect to rootclient page with error message
 };
 
-module.exports = mobileController;
+module.exports = clientController;
